+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog].
 - Remove Zero Sized Polygons `#659`
 - Add support for UniVRM components `#653`
 - Support for Mesh Topologies other than Triangles `#692`
+- Skip enablement mismatched Renderers in Merge Skinned Mesh `#670`
+  - This is now enabled by default for newly added Merge Skinned Mesh.
 
 ### Changed
 - All logs passed to ErrorReport is now shown on the console log `#643`